Remove -fno-exceptions -fno-rtti
Adds summary of loops/sec to CursesRenderer.

Using this stat, in my crude testing, showed that the -fno-rtti and
-fno-exceptions optimizations were actually HURTING performance.

With either or both build options, the total loops/sec calculated is about
15,800 around 10 seconds after program start* (on an 8 virtual-core i7-4790K @
4.00GHz).

With neither given, performance jumps to 16,000 with an identical test.

Performance drops as lifeform Dna grows, so 10 seconds after start is a
compromise between startup sputter and this.  The numbers above are stable
across multiple executions in random order.

Therefore, we remove these options from the Makefile.